pyproject.toml
[project]
name = "obsidian-sync"
version = "0.1.0"
description = "Git sync daemon for Obsidian vaults"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
obsidian-sync = "obsidian_sync.cli:main"

[project.urls]
Homepage = "https://github.com/safurrier/obsidian-sync"
Repository = "https://github.com/safurrier/obsidian-sync"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-xdist>=3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_sync"]

[tool.hatch.build.targets.wheel.force-include]
"src/obsidian_sync/templates" = "obsidian_sync/templates"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
